Add files via upload
This is a web development project in a 22 week software dev intensive course with NuCamp.  This website is used building Bootstrap and some jquery.  Server connection isn't hooked up at this point, but there is a login page, a reservation button that links to a form, a carousel/tabs, and the different html pages link to each other.  Node modules file is missing, but will be needed to run the program.